+++ src/main/java/com/intellecteu/onesource/integration/services/client/onesource/dto/RecallDTO.java
@@ -14,7 +14,6 @@ package com.intellecteu.onesource.integration.services.client.onesource.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
@@ -38,10 +37,10 @@ public class RecallDTO {
   private VenueDTO executionVenue = null;
 
   @JsonProperty("openQuantity")
-  private BigDecimal openQuantity = null;
+  private Integer openQuantity = null;
 
   @JsonProperty("quantity")
-  private BigDecimal quantity = null;
+  private Integer quantity = null;
 
   @JsonProperty("recallDate")
   private LocalDate recallDate = null;
@@ -124,7 +123,7 @@ public class RecallDTO {
     this.executionVenue = executionVenue;
   }
 
-  public RecallDTO openQuantity(BigDecimal openQuantity) {
+  public RecallDTO openQuantity(Integer openQuantity) {
     this.openQuantity = openQuantity;
     return this;
   }
@@ -134,15 +133,15 @@ public class RecallDTO {
    * @return openQuantity
   **/
   @Schema(description = "")
-  public BigDecimal getOpenQuantity() {
+  public Integer getOpenQuantity() {
     return openQuantity;
   }
 
-  public void setOpenQuantity(BigDecimal openQuantity) {
+  public void setOpenQuantity(Integer openQuantity) {
     this.openQuantity = openQuantity;
   }
 
-  public RecallDTO quantity(BigDecimal quantity) {
+  public RecallDTO quantity(Integer quantity) {
     this.quantity = quantity;
     return this;
   }
@@ -152,11 +151,11 @@ public class RecallDTO {
    * @return quantity
   **/
   @Schema(required = true, description = "")
-  public BigDecimal getQuantity() {
+  public Integer getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(BigDecimal quantity) {
+  public void setQuantity(Integer quantity) {
     this.quantity = quantity;
   }
