+++ src/main/java/com/intellecteu/onesource/integration/services/client/onesource/dto/ReturnProposalDTO.java
@@ -14,7 +14,6 @@ package com.intellecteu.onesource.integration.services.client.onesource.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 /**
@@ -28,7 +27,7 @@ public class ReturnProposalDTO {
   private VenueDTO executionVenue = null;
 
   @JsonProperty("quantity")
-  private BigDecimal quantity = null;
+  private Integer quantity = null;
 
   @JsonProperty("returnDate")
   private LocalDate returnDate = null;
@@ -63,7 +62,7 @@ public class ReturnProposalDTO {
     this.executionVenue = executionVenue;
   }
 
-  public ReturnProposalDTO quantity(BigDecimal quantity) {
+  public ReturnProposalDTO quantity(Integer quantity) {
     this.quantity = quantity;
     return this;
   }
@@ -73,11 +72,11 @@ public class ReturnProposalDTO {
    * @return quantity
   **/
   @Schema(required = true, description = "")
-  public BigDecimal getQuantity() {
+  public Integer getQuantity() {
     return quantity;
   }
 
-  public void setQuantity(BigDecimal quantity) {
+  public void setQuantity(Integer quantity) {
     this.quantity = quantity;
   }
